refactor: extract getStatusByDate
