+++ client/src/modules/AutoTest/components/TaskDeadlineDate/TaskDeadlineDate.tsx
@@ -1,9 +1,9 @@
 import { Space, Typography } from 'antd';
 import { CalendarOutlined } from '@ant-design/icons';
-import React from 'react';
-import moment from 'moment';
+import React, { useMemo } from 'react';
 import { dateWithTimeZoneRenderer } from 'components/Table';
 import { BaseType } from 'antd/lib/typography/Base';
+import getStatusByDate, { AutoTestTaskStatus } from 'modules/AutoTest/utils/getStatusByDate';
 
 const { Text } = Typography;
 
@@ -13,36 +13,30 @@ type TaskDeadlineDateProps = {
   score?: number;
 };
 
+function getTextType(endDate: string, score: number): BaseType {
+  const status = getStatusByDate(endDate, score);
+
+  switch (status) {
+    case AutoTestTaskStatus.Missed:
+      return 'danger';
+    default:
+      return 'secondary';
+  }
+}
+
 function TaskDeadlineDate({ startDate, endDate, score = 0 }: TaskDeadlineDateProps): JSX.Element {
   const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-  function getTextType(): BaseType {
-    const now = moment();
-    const end = moment(endDate);
-
-    const isDeadlineSoon = now <= end && end.diff(now, 'hours') < 48 && !score;
-    const isDeadlineMissed = now >= end && !score;
-
-    if (isDeadlineMissed) {
-      return 'danger';
-    }
-
-    if (isDeadlineSoon) {
-      return 'warning';
-    }
-
-    return 'secondary';
-  }
+  const type = useMemo(() => getTextType(endDate, score), [endDate, score]);
 
   const start = dateWithTimeZoneRenderer(timezone, 'MMM DD')(startDate);
   const end = dateWithTimeZoneRenderer(timezone, 'MMM DD')(endDate);
 
   return (
     <Space wrap style={{ justifyContent: 'end' }}>
-      <Text type={getTextType()}>
+      <Text type={type}>
         <CalendarOutlined />
       </Text>
-      <Text type={getTextType()}>
+      <Text type={type}>
         {start} &ndash; {end}
       </Text>
     </Space>
