+++ client/src/modules/AutoTest/utils/parseCourseTask.ts
